Revert "cirrus: Send more_like traffic to eqiad"
This reverts commit 887d06acf2a5c1dce6c31c5a9a6e6f887e24e272.

Reason for revert:

requestctl rules have been setup to stop the offending traffic,
the search cluster is now healthy and ready for the normal traffic
load.

Additional classifications are also being added to CirrusSearch,
but not deployed yet, that will attempt to further prevent this
kind of overload from automated requests in the future.

Bug: T405394
Change-Id: I0353a57c4615deeb48e7138b9b7d8eec7a97645a