Disable inert MobileFrontend on wikimedia.org wikis that lack DNS
These wikis don't have mobile DNS, don't have a Varnish mobile_redirect,
and so for all intents and purposes have no mobile version.

The only sign of it being installed is Special:Version and a broken
"Mobile view" footer link, to a non-existent domain name.

To avoid surprises, disable MobileFrontend here ahead of rolling out
unified mobile routing (T403510) as otherwise we would effectively
enable MobileFrontend on these wikis for the first time, which is out of
scope for this project. Outreach and audits is especially hard for these
either private or closed wikis.

Bug: T152882
Change-Id: Ic4bcaf0f2a63649ece56d6f45834f24022688647