Enable REST Sandbox on all wikis
The REST Sandbox special page was initially enabled on only a small
set of production wikis, for testing and feedback. That initial
time period is now complete, and all necessary fixes and adjustments
have been completed for wider availablity of the REST Sandbox. So
enable it everywhere, with an initial set of specs. Additional specs
(such as endpoints in extensions) can be enabled later in followup
changes if desired.

Bug: T389409
Change-Id: I8c808b6df4b13d835b7620573567da9470718e61