Enable the MEX / wbui2025 beta feature on testwikidata
In T394704 (Ib76c441cafd16) we developed a beta feature flag to
support users opting in to the new / experimental wbui2025 rewrite
of the Wikidata item editing interface for mobile devices.

Enable the beta feature flag on testwikidata so that logged-in
users can turn the feature on. Leave the flag disabled on beta
and wikidata so that no other users see the beta feature.

Group0-Depends-On: I5ebf9743642fafe14b08e2dddbc1d29c90e5576c
Bug: T407737
Change-Id: I400b6c3eaac71e961f7f6867d84ead98c14e5e3d