Remove nlwiki exception from thumb limits
We explicitly don't allow exceptions on thumb sizes anymore specially on
the default thumb size (which effectively this exception bumps the
default from 250px to 330px on nlwiki) which causes cache fragmentation
and slow down of loading of images for everyone (specially in northern
Europe given where esams is). The change is basically invisible to the
naked eye.

Bug: T408715
Change-Id: Ib23ab40b97ef7930e35415faad15bd473e321aea