Reduce number of bucketsizes for MediaViewer (labs, group0)
MediaViewer assumes that certain thumbnail sizes are more likely to
have been previously rendered than others, and will lock to the given
bucket sizes when loading up an image to fit the current window size.

The default set of buckets is relatively large and meant to cover
many screen sizes, but this produces some fragmentation across
devices and window sizes, making it less likely that any given image
will have been pre-rendered and saving multiple output versions
to disk.

As an optimization to reduce extra thumbnail churn, we're reducing
it to a smaller set:

    `[ 400, 960, 1280, 2560 ]`

meaning that fewer intermediate resolutions will be chosen on
small and middle-sized screens.

Sets the updated config for all beta wikis and group0 production
(includes test.wikipedia.org). If no problems encountered, we'll
push the default shortly.

Expected results:

* unchanged behavior and performance for most users
* over time, reduced latency for MediaViewer loads is expected
* fewer stray small thumbnails will be created, saving disk space
* possible changes to bandwidth usage as (increase from fewer
  small versions, decrease from fewer large versions); unlikely
  to be measurable

Note a previous version of this config patch was reverted due to a
config merge setup problem, which is resolved in MultimediaViewer
in 8fa2ce06.

Recommended follow-up work:
* look at wgThumbnailSteps and wgUploadThumbnailRenderMap to ensure
  that the default thumbnail sizes in production and MediaViewer
  are prerendered on upload

Bug: T372165
Change-Id: I029521e7450f4f16a7f202c2a2b0a005dab93004