Reduced MediaViewer bucket sizes list to group1
Extends the rollout of the reduced MediaViewer bucket sizes list
to group1 (previously on all beta & main group0).

This reduces the number of different sizes that the big zoom view
for MediaViewer will try to create, making it more likely that
we'll get a previously-rendered hit which willl load faster.

See also future discussion on T408062 for standardizing these
bucket sizes later.

Bug: T372165
Change-Id: If4b6ff2f767d56cbb130df36156bae55395696e3