tcywikisource: Migrate $wgAccountCreationThrottle to throttle.php
Apparently the IP/range in throttle.php is optional (though ThrottleTest
needs to be taught to avoid “This test did not perform any assertions”
errors), and this way the exception can be limited to a tighter time
range (though extended by one hour past the scheduled end of the event
in case they overrun a bit).

Also remove one expired exception.

Bug: T410507
Change-Id: Ibb4856125ad1f7d405b2866476d06f564c70d52e