testwiki: enable ReviseTone experiment
We need to override the search term because the test-articles for this
feature have all been newly generated in the last weeks, but in
production we are excluding all articles created in the last 90 days.

The time for `lasteditdate` is to "during the previous hour or older"
to allow for tone suggestions to show up again faster on testwiki. We
will likely only have a very small number of suggestions there, so we
need to be economical about them. On production this value will be
unmodified and be "older than the last 24 previous hours".

Bug: T407029
Change-Id: I0e0dfdddfa6a10b16b0ea0c8268575d35598b852