[mediawikiwiki] Enable CommunityRequests with translations only
This is so that we can use {{int:}} on our messages in the documentation
and possibly also be able to transclude the Special:ApiHelp docs.
Err on the side of caution by not touching $wgVirtualDomainsMapping when
we know we aren't going to use it, even though it may not actually make
a difference one way or another.

Also change log level to info. We do not need to actively monitor the
debug messages anymore.

Bug: T405694
Change-Id: Ie59f0f1c52c5ac00ee27a6c45a3a86af909cee4a