robots.txt: Remove redundant "/wiki/Fundraising_2007/comments" disallow
This was added in 2007 to the global robots.txt because at that time
we did not yet have per-wiki [[MediaWiki:Robots.txt]] or per-page
`__NOINDEX__` in the Parser, which was added to MediaWiki in 2008.

I've edited these pages on Meta-Wiki to include __NOINDEX__, so this
global rule is no longer needed.

https://meta.wikimedia.org/wiki/Fundraising_2007/comments/allpages

Change-Id: Ibf0daf97500553736f89f24449086938f904317b