Enable the MEX / wbui2025 beta feature on wikidata
Following-up on I60cd21e80c47ec2, we are now ready to enable the MEX
user interface on wikidata.org production under a beta feature flag.

Update the deployment config to make the feature available to users.
`wikibase-mobile-editing-ui` is already included in the
`wgBetaFeaturesAllowList` for the wikidata dblist, so this should
be as simple as setting `wmgWikibaseTmpMobileEditingUIBetaFeature` to
`true` on wikidata.org.

Additional changes to the copy for the beta feature will follow in
T411341.

Bug: T403015
Change-Id: Iedec39a923a311f9b75359edbe325813ecb43432