Update fc-list to point to fc-list Tool
Why:
fc-list Tool, hosted on Toolforge is running now daily, updating the
corresponding list of fonts using the information from the
deployment-charts repo to target the most recently deployed Thumbor OCI
image. It is going to be far, far closer to reality than this hand
maintained list ever was

What:
Replace the content with a mention and a link to the Toolforge hosted
fc-list tool. Given the amount of users and the fact that this file
never really belonged in this repo, I opted for not implementing any
HTTP or HTML redirects. Users are nudged towards updating their
bookmarks and habbits

Bug: T280718
Change-Id: Ieef0d2f8125b92afa97455f7fa5a15bfbf6c48e5